% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_helper.R
\name{cross_fit_helper}
\alias{cross_fit_helper}
\title{Helper function to for cross-fitting}
\usage{
cross_fit_helper(model_W, model_Y, folds_to_fit, use)
}
\arguments{
\item{model_W}{The model used to produce the estimates of W_hat}

\item{model_Y}{The model used to produce the estimates of Y_hat}

\item{folds_to_fit}{The split folds for which the calculations are being run.}

\item{use}{The method to use when calculating the out of fold prediction (propagates from method).}
}
\description{
Fits estimated models to out of fold samples
}
\examples{

\dontrun{
 cross_fit_helper( model_W = W_fit_model,
                   model_Y = Y_fit_model,
                   folds_to_fit = other_folds,
                   use = "glmnet")

}
}
