% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ML.R
\name{plot_ML}
\alias{plot_ML}
\title{Plot the fitted Debiased ML treatment effects.}
\usage{
plot_ML(x, coloured.same = TRUE, palette = NULL, ...)
}
\arguments{
\item{x}{The estimated ML_Treatment_effects object to plot}

\item{coloured.same}{Whether to use the same colour for all plots, defaults to TRUE.}

\item{palette}{An optional custom colour/vector of palette colour codes to be used for plotting.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
A function to plot the estimated treatment effects.
}
\details{
The plot samples a random colour from a predefined colourblind palette by default.
You may use a custom colour/colour palette for plotting by supplying the palette argument.
If you supply a palette, the colour will be used in the order in which they are specified.
Palette lengths longer than the number of plots will be ignored.
}
\examples{
\dontrun{

    ols_example <- double_ML(X, Y, W, method = c("ols"),
                             show.progress = FALSE,
                             k.fld = 4,
                             simulations = 50)

   plot_ML(ols_example)




}

}
